+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## [Unreleased]
 
+- Added type/object and type/input-object
 - Improved logging
-- Improved object type introspection
+- Improved object and list type introspection
 - Use Clojure 1.10.0
 
 ## [0.5.0] - 2018-07-31